# Reduce latency effects of non-interactive I/O.