@amotin 2020-11-07
@IvanVolosyuk In theory I think it is possible, but I was unable to reproduce it when tried. It has to be perfectly sequential data without holes, the reader should not distract on indirect blocks and other metadata and it should never lag.