@amotin 2020-11-07
I've updated the patch following most of @behlendorf comments.