@h1z1 2020-11-10
Curious about this, what kind of IO were you submitting and was readahead disabled?  I believe scrub IO is async no?  Those drives have quite large caches.