@amotin 2020-11-10
@h1z1 As I have written, payload was 4KB random read, scrub was sequential.  Disk settings were at defaults -- both read-ahead and write cache enabled.  Not sure what you mean mentioning async and cache.