@h1z1 2020-11-10
> @h1z1 As I have written, payload was 4KB random read, scrub was sequential. Disk settings were at defaults -- both read-ahead and write cache enabled. Not sure what you mean mentioning async and cache.

> both read-ahead and write cache enabled.

Good to know, I didn't see that above?  Was readahead also disabled otherwise your reads will be amplified no?

> Not sure what you mean mentioning async and cache.

They are async as in they are not expected to return with any set deadline.  You can end up in cases where the drive has the data you're about to request in cache but evicts it before your submission.  Or it could optimize the operations.   I'd expect drives to starve random IO if they're getting blasted with sequential especially it doesn't overlap.