@amotin 2020-11-10
> What about OS?

FreeBSD.

> > both read-ahead and write cache enabled.
> 
> Good to know, I didn't see that above? Was readahead also disabled otherwise your reads will be amplified no?

As I have told, everything is at disk defaults.  FreeBSD does not change them unless asked.  Readahead is enabled by default.

> I'd expect drives to starve random IO if they're getting blasted with sequential especially it doesn't overlap.

I am not sure queue depth of 1 counts as blasted. I clearly understand why the drive would behave the way it does, it is just very far from fail scheduling.