@amotin 2020-11-10
@ahrens Before me scrub was always using queue depth of 2. I've made it jump between 1 and 3 (which is more safe now).  So despite the min value changed, I don't think it is a step down.  I would increase max even higher, unless on FreeBSD due to MAXPHYS of 128KB and ZFS I/O aggregation to 1MB it would not stuff SATA queue completely full already at 4.