@h1z1 2020-11-11
> > What about OS?
> 
> FreeBSD.
> 

o_O And what about Linux?

> > > both read-ahead and write cache enabled.
> > 
> > 
> > Good to know, I didn't see that above? Was readahead also disabled otherwise your reads will be amplified no?
> 
> As I have told, everything is at disk defaults. FreeBSD does not change them unless asked. Readahead is enabled by default.
> 

You haven't told me anything until I asked and that doesn't explain what the defaults are.

> > I'd expect drives to starve random IO if they're getting blasted with sequential especially it doesn't overlap.
> 
> I am not sure queue depth of 1 counts as blasted. I clearly understand why the drive would behave the way it does, it is just very far from fail scheduling.

I don't think you do tbh given you're completely ignoring the amplification as a cause nor said what IOPs you're testing with. And if the IO is async does FreeBSD still limit the queue to 1 -outstanding- request?