@amotin 2020-11-11
> o_O And what about Linux?

And what about Linux?  Is it special?

> You haven't told me anything until I asked and that doesn't explain what the defaults are.

"Disk settings were at defaults -- both read-ahead and write cache enabled."  Could I be more specific anyhow?

> I don't think you do tbh given you're completely ignoring the amplification as a cause nor said what IOPs you're testing with. 

Would you be more specific in what ignorance exactly are you accusing me?  Which of amplifications are you talking about?  "I've benchmarked this change with 4KB random reads from ZVOL with 16KB block size on newly written non-fragmented pool."

> And if the IO is async does FreeBSD still limit the queue to 1 -outstanding- request?

No.  FreeBSD by default does not do any I/O scheduling, passing everything to the drive as much as it fits.