@amotin 2020-11-11
> I don't know, what's so special about FBSD? Did you test it at all on Linux?

No, I didn't.  Are you arguing just to argue, or you have some reason to think it is different on Linux?  If you want to help with testing -- please, be my guest.

> .. and you're still not saying what your IOPs were nor even what you used to test with. fio? dd? Some custom util? You're using a 16k blocksize against the zvol or device itself??

Who cares what the tool was if it was simple QD1 4KB random read?  If you are curios, it was `fio` with such a config:
`[global]
size=100G
runtime=500
ioengine=psync
iomem_align=2m
blocksize=4k
rw=randread
iodepth=1
zero_buffers
group_reporting

[job]
filename=/dev/zvol/small/vol1
write_hist_log=lsmall
log_hist_msec=60000
log_hist_coarseness=6`

If you need average test IOPS during scrub, then for SEAGATE ST14000NM0048 it increased from 22 to 73, without scrub it was 166.

16KB is a volblocksize. The disk is 512e, and pool has ashift=12.

> .. to be clear, you're telling me FreeBSD has no IO scheduler?

There are some specialized, but by default there is nothing other then FIFO (for SSDs) or elevator (for HDDs) for requests that don't fit the device cache.  We already have one scheduler in disk, and another inside ZFS. How many more schedulers do we need?